Create PendingApproval class
